+++ src/main/kotlin/cn/yiiguxing/plugin/translate/trans/TranslationNotifications.kt
@@ -3,11 +3,19 @@ package cn.yiiguxing.plugin.translate.trans
 import cn.yiiguxing.plugin.translate.action.TranslationEngineActionGroup
 import cn.yiiguxing.plugin.translate.message
 import cn.yiiguxing.plugin.translate.util.Notifications
+import cn.yiiguxing.plugin.translate.util.concurrent.finishOnUiThread
+import cn.yiiguxing.plugin.translate.util.concurrent.successOnUiThread
 import cn.yiiguxing.plugin.translate.util.e
+import com.intellij.ide.DataManager
 import com.intellij.notification.Notification
 import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.actionSystem.PlatformDataKeys
+import com.intellij.openapi.actionSystem.PopupAction
+import com.intellij.openapi.application.ModalityState
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.project.Project
+import org.jetbrains.concurrency.runAsync
 import java.util.*
 
 object TranslationNotifications {
@@ -34,7 +42,7 @@ object TranslationNotifications {
         val actionList = LinkedList<AnAction>()
         errorInfo?.continueActions?.let { actionList += it }
         actionList.addAll(actions)
-        actionList.add(TranslationEngineActionGroup({ message("action.SwitchTranslationEngineAction.text") }))
+        actionList.add(SwitchTranslationEngineAction())
 
         if (throwable !is TranslateException) {
             // 将异常写入IDE异常池，以便用户反馈
@@ -46,4 +54,26 @@ object TranslationNotifications {
         }
     }
 
+    private class SwitchTranslationEngineAction : AnAction({ message("action.SwitchTranslationEngineAction.text") }),
+        PopupAction {
+
+        private var isActionPerforming = false
+
+        override fun actionPerformed(e: AnActionEvent) {
+            val component = PlatformDataKeys.CONTEXT_COMPONENT.getData(e.dataContext) ?: return
+            if (isActionPerforming) {
+                return
+            }
+            isActionPerforming = true
+            runAsync { TranslationEngineActionGroup() }
+                .successOnUiThread { group ->
+                    if (component.isShowing) {
+                        // Do not use `e.dataContext` directly because it is not an async data context.
+                        val dataContext = DataManager.getInstance().getDataContext(component)
+                        group.showActionPopup(dataContext)
+                    }
+                }
+                .finishOnUiThread(ModalityState.any()) { isActionPerforming = false }
+        }
+    }
 }