+++ src/main/kotlin/cn/yiiguxing/plugin/translate/ui/TranslationBalloon.kt
@@ -81,6 +81,7 @@ class TranslationBalloon(
         Disposer.register(balloon) { Disposer.dispose(this) }
         Disposer.register(this, processPane)
         Disposer.register(this, translationPane)
+        Disposer.register(this, errorPanel)
 
         ApplicationManager
             .getApplication()
@@ -255,7 +256,6 @@ class TranslationBalloon(
     companion object {
 
         private const val MAX_WIDTH = 500
-        private const val MIN_ERROR_PANEL_WIDTH = 300
         private const val INSETS = 20
 
         private const val CARD_PROCESSING = "processing"